Done Post multiple Posts, Delete
The user can post multiple blot posts and can delete them if needed. Furthermore, the created blog is presented further on a separate site if needed by the user. Need to fix: the update button has issues. Replacing the current CSS with a more responsive one. The blog posts are presented cleanly in the home page.